requirements.txt
numpy==2.2.5
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
pytesseract==0.3.13
